Finalize Topbar component
