+++ packages/components/src/community/index.tsx
@@ -1,2 +1,4 @@
 export { CHANNEL_GROUPS } from './mock-data'
 export { SidebarCommunity } from './sidebar-community'
+export { SidebarMembers } from './sidebar-members'
+export { Topbar } from './topbar'
