+++ packages/components/src/index.tsx
@@ -1,5 +1,7 @@
 export * from './anchor-actions'
+export * from './avatar'
 export * from './button'
+export * from './community'
 export * from './composer'
 export * from './dividers'
 export * from './dynamic-button'
@@ -11,16 +13,10 @@ export * from './input'
 export * from './messages'
 export * from './pinned-message'
 export * from './provider'
-export * from './sidebar'
-export * from './sidebar-members'
 export * from './skeleton'
 export * from './text'
 export * from './toast'
-export * from './topbar'
 export * from './user-list'
-
-// MOCK DATA
-export { CHANNEL_GROUPS } from './sidebar/mock-data'
 
 // eslint-disable-next-line simple-import-sort/exports
 export { config } from './tamagui.config'
